refactor: Move Visualization component
